+++ src/Features/ERDDAP/hooks/buoyBarn.ts
@@ -1,5 +1,5 @@
 import * as Sentry from "@sentry/react"
-import { useQuery } from "react-query"
+import { useQuery, useQueries } from "react-query"
 
 import { ForecastJson, ForecastSource, PlatformFeatureCollection } from "../types"
 import { defaultQueryConfig } from "./hookConfig"
@@ -54,7 +54,7 @@ const getForecasts = async () => {
 /**
  * Load forecasts
  */
-export function useForecasts() {
+export function useForecastMeta() {
   return useQuery("buoybarn-forecasts", getForecasts, defaultQueryConfig)
 }
 
@@ -95,4 +95,21 @@ export function useForecast(lat: number, lon: number, forecast?: ForecastSource)
     ...defaultQueryConfig,
     enabled: forecast?.source_url ? true : false,
   })
+}
+
+/**
+ * Load multiple forecasts for a given point
+ *
+ * @param lat Latitude in decimal degrees
+ * @param lon Logitude in decimal degrees
+ * @param forecasts Forecast sources to load
+ */
+export function useForecasts(lat: number, lon: number, forecasts: ForecastSource[]) {
+  return useQueries(
+    forecasts.map((forecast) => ({
+      ...defaultQueryConfig,
+      queryKey: ["buoybarn-forecast", { forecast, lat, lon }],
+      queryFn: () => getForecast(forecast, lat, lon),
+    }))
+  )
 }